Add scripts with libtcl86 preload to access db4 via tkinter on OpenBSD.
This amendment affects development environment only.  Future changes may
be forced by changes to /usr/local/lib/libtcl<x>.so.<y> version numbers.

It is possible to install db4 from ports, with some minor amendments to
the makefile, so preload is not necessary.

This is a workaround for loss of access to current package versions of
Berkeley DB via either berkeleydb or bsddb3 at Python 10 on OpenBSD.